refactor(log): filter 이름 변경
